Moved EC2 repository to proper package
