+++ src/main/java/org/nephology/IndexController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
  * Created by PLGrubskiM on 2017-02-12.
  */
 @Controller
-public class HelloController {
+public class IndexController {
 
     @RequestMapping(value = "/")
     public String index() {
