+++ src/main/java/org/nephology/aws/ec2/domain/AwsEC2InstanceDetailsDataRepository.java
@@ -1,6 +1,5 @@
-package org.nephology;
+package org.nephology.aws.ec2.domain;
 
-import org.nephology.aws.ec2.domain.AwsEC2InstanceDetailsData;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface AwsEC2InstanceDetailsDataRepository extends MongoRepository<AwsEC2InstanceDetailsData, String> {
